refactor: consolidate connection configuration into RelayConfig
- Move ConnectionConfig validation into RelayConfig
- Simplify error types by removing specific variants
- Update HTTP server configuration with bind_addr
- Change RTU max_frame_size type to u64
- Add RTS configuration options
- Improve duration handling in config serialization
- Move logging validation to RelayConfig
- Add additional validation for per_ip_limits

This change improves configuration management by:
1. Centralizing all validation logic in RelayConfig
2. Using consistent types and formats for configuration values
3. Removing redundant error types and validation code
4. Adding missing configuration options

+++ src/config/backoff.rs
@@ -1,15 +1,19 @@
 use std::time::Duration;
 
-#[derive(Debug, Clone)]
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Config {
     /// Initial wait time
+    #[serde(with = "humantime_serde")]
     pub initial_interval: Duration,
     /// Maximum wait time
+    #[serde(with = "humantime_serde")]
     pub max_interval: Duration,
     /// Multiplier for each subsequent attempt
     pub multiplier: f64,
     /// Maximum number of attempts
-    pub max_retries: usize,
+    pub max_retries: u64,
 }
 
 impl Default for Config {
